feat: interactive sites list explorer — filters, presets, url sync, export
