+++ src/app/sites-list/page.tsx
@@ -3,27 +3,9 @@ import HealthDashboard from './HealthDashboard'
 import DomainExpiry from './DomainExpiry'
 import NonprofitCallout from '@/components/NonprofitCallout'
 import { loadDomainExpiry, relativeAge } from '@/lib/dashboardData'
-import { isGithubHandle } from '@/data/site-owners'
 import { ViewNav } from './PersonaView'
-import CopyRowButton from './CopyRowButton'
-import {
-  SiteData,
-  loadSites,
-  dataRefreshedAge,
-  healthBadge,
-  healthCategory,
-  cloudflareZoneUrl,
-  cloudflareDnsRecordsUrl,
-  dnsCheckerUrl,
-  pagesSettingsUrl,
-  newIssueUrl,
-  rowMarkdown,
-  migrationSteps,
-  sslBadge,
-  lighthouseBadge,
-  dataGeneratedAt,
-  dataIsStale,
-} from './sitesData'
+import SitesExplorer from './SitesExplorer'
+import { loadSites, dataRefreshedAge, dataGeneratedAt, dataIsStale } from './sitesData'
 
 export const metadata: Metadata = {
   title: 'Sites Master List',
@@ -31,352 +13,12 @@ export const metadata: Metadata = {
     'Volunteer-focused dashboard of FFC-managed domains, grouped by how much effort they need — active development, stalled, needs migration, done, triage, and inactive.',
 }
 
-const TIERS = [
-  {
-    num: '1',
-    icon: '🔨',
-    title: 'Active Development',
-    blurb:
-      'A GitHub repo with a pull request closed in the last 45 days. Momentum is here — the fastest way to help is to keep these moving.',
-    header: 'bg-green-100 text-green-900 border-green-200',
-    open: true,
-  },
-  {
-    num: '2',
-    icon: '🌱',
-    title: 'Has Repo, Stalled',
-    blurb:
-      'A repo exists but has gone quiet (no recent merged PRs). High-value to revive — the groundwork is already done.',
-    header: 'bg-yellow-100 text-yellow-900 border-yellow-200',
-    open: true,
-  },
-  {
-    num: '3',
-    icon: '🚚',
-    title: 'Needs Migration',
-    blurb:
-      'Active domains still served from legacy hosting (HostPapa, InterServer, Hostinger, Krystal). The goal is to move them to a GitHub Pages repo.',
-    header: 'bg-blue-100 text-blue-900 border-blue-200',
-    open: true,
-  },
-  {
-    num: '4',
-    icon: '✅',
-    title: 'Done / Stable',
-    blurb:
-      'Fully migrated (apex domain + Cloudflare + GitHub Pages) and responding. The desired end state — little effort needed.',
-    header: 'bg-teal-100 text-teal-900 border-teal-200',
-    open: false,
-  },
-  {
-    num: '5',
-    icon: '🔍',
-    title: 'Needs Triage',
-    blurb:
-      'Active domains with no matched repo and unclear hosting. Investigate (is there a site? a repo elsewhere?) before assigning volunteer effort.',
-    header: 'bg-gray-100 text-gray-900 border-gray-200',
-    open: false,
-  },
-  {
-    num: '6',
-    icon: '💤',
-    title: 'Inactive / Archive',
-    blurb:
-      'Expired, cancelled, fraud, terminated, or transferred away. Not worth volunteer effort.',
-    header: 'bg-red-100 text-red-900 border-red-200',
-    open: false,
-  },
-]
-
-function tierRank(s: SiteData): number {
-  return { live: 1, redirect: 2, error: 3, unreachable: 4, unknown: 5 }[
-    healthCategory(s.siteHealth)
-  ]
-}
-
-// Within a tier, lead with the most recently active (tiers 1/2) or healthiest.
-function sortTier(sites: SiteData[], num: string): SiteData[] {
-  return [...sites].sort((a, b) => {
-    if (num === '1' || num === '2') {
-      if (a.lastPrClosed !== b.lastPrClosed) return b.lastPrClosed.localeCompare(a.lastPrClosed)
-    }
-    const r = tierRank(a) - tierRank(b)
-    if (r !== 0) return r
-    return a.domain.localeCompare(b.domain)
-  })
-}
-
-function repoName(url: string): string {
-  return url.replace('https://github.com/', '')
-}
-
-// Data source per column, surfaced as header tooltips and in the legend (#417).
-const COLUMN_SOURCE: Record<string, string> = {
-  Domain: 'WHMCS export (domain inventory)',
-  Repo: 'GitHub API',
-  'Last PR': 'GitHub API',
-  'Open PRs': 'GitHub API',
-  Health: 'HTTP health probe',
-  SSL: 'TLS probe (upstream pipeline)',
-  LH: 'Lighthouse CI (upstream pipeline)',
-  Owner: 'src/data/site-owners.ts in this repo',
-  Progress: 'Derived from Cloudflare, repo, server, and health fields',
-  Server: 'HTTP health probe / host detection',
-  Status: 'WHMCS export',
-  Notes: 'WHMCS export',
-}
-
-// Compact per-row quick-action links under the domain (#408, #410, #411, #423).
-function RowActions({ s }: { s: SiteData }) {
-  const action = (href: string, label: string, title: string) => (
-    <a
-      key={label}
-      href={href}
-      target="_blank"
-      rel="noopener noreferrer"
-      title={title}
-      aria-label={title}
-      className="text-gray-500 hover:text-gray-800 hover:underline"
-    >
-      {label}
-    </a>
-  )
-  const cf = cloudflareZoneUrl(s)
-  const dnsRecords = cloudflareDnsRecordsUrl(s)
-  const actions = [
-    cf && action(cf, '☁ CF', `Open ${s.domain} in the Cloudflare dashboard`),
-    dnsRecords && action(dnsRecords, 'DNS', `Cloudflare DNS records for ${s.domain}`),
-    action(dnsCheckerUrl(s.domain), 'check', `DNS propagation check for ${s.domain}`),
-    s.repoUrl &&
-      action(
-        s.repoUrl,
-        s.openPrs && s.openPrs !== '0' ? `repo (${s.openPrs} PRs)` : 'repo',
-        `GitHub repository: ${repoName(s.repoUrl)}`
-      ),
-    pagesSettingsUrl(s) &&
-      action(pagesSettingsUrl(s), 'pages', `GitHub Pages settings for ${s.domain}`),
-    action(newIssueUrl(s), '+issue', `Open a prefilled GitHub issue about ${s.domain}`),
-  ].filter(Boolean)
-  return (
-    <div className="mt-0.5 flex flex-wrap gap-x-2 gap-y-0.5 text-[11px] leading-4">
-      {actions}
-      <CopyRowButton text={rowMarkdown(s)} domain={s.domain} />
-    </div>
-  )
-}
-
-// 4-step migration progress dots, derived from existing fields (#424).
-function MigrationProgress({ s }: { s: SiteData }) {
-  const steps = migrationSteps(s)
-  const done = steps.filter((x) => x.done).length
-  return (
-    <span title={steps.map((x) => `${x.done ? '✓' : '○'} ${x.label}`).join('\n')}>
-      <span aria-hidden="true" className="tracking-widest text-ffc-teal-dark">
-        {steps.map((x) => (x.done ? '●' : '○')).join('')}
-      </span>
-      <span className="sr-only">{`${done} of ${steps.length} migration steps complete`}</span>
-      <span className="ml-1 text-[11px] text-gray-500 tabular-nums">{done}/4</span>
-    </span>
-  )
-}
-
-function TierTable({ sites, num }: { sites: SiteData[]; num: string }) {
-  // Dev tiers (1, 2) lead with repo activity; other tiers hide those columns.
-  const showRepoCols = num === '1' || num === '2'
-  // Optional data-signal columns render only when the snapshot carries data.
-  const hasSsl = sites.some((s) => s.sslExpiry)
-  const hasLh = sites.some((s) => s.lighthouse)
-  const hasOwner = sites.some((s) => s.owner)
-  const showProgress = num === '3'
-  const headers = [
-    'Domain',
-    ...(showRepoCols ? ['Repo', 'Last PR', 'Open PRs'] : []),
-    'Health',
-    ...(hasSsl ? ['SSL'] : []),
-    ...(hasLh ? ['LH'] : []),
-    ...(hasOwner ? ['Owner'] : []),
-    ...(showProgress ? ['Progress'] : []),
-    'Server',
-    'Status',
-    'Notes',
-  ]
-  const dash = <span className="text-gray-300">—</span>
-  return (
-    <div className="overflow-x-auto">
-      <table className="min-w-full divide-y divide-gray-200 text-sm">
-        <thead className="bg-gray-50">
-          <tr>
-            {headers.map((h) => (
-              <th
-                key={h}
-                scope="col"
-                title={COLUMN_SOURCE[h] ? `Source: ${COLUMN_SOURCE[h]}` : undefined}
-                className="px-4 py-2 text-left text-xs font-bold uppercase tracking-wider text-gray-500"
-              >
-                {h}
-                {COLUMN_SOURCE[h] && (
-                  <span className="sr-only">{` (source: ${COLUMN_SOURCE[h]})`}</span>
-                )}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody className="bg-white divide-y divide-gray-100">
-          {sites.map((s) => {
-            return (
-              <tr key={`${num}-${s.domain}`} className="hover:bg-gray-50">
-                <td className="px-4 py-2 whitespace-nowrap font-medium">
-                  <div className="flex items-center gap-1.5">
-                    <a
-                      href={`https://${s.domain}`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-blue-600 hover:underline"
-                    >
-                      {s.domain}
-                    </a>
-                    {s.nsMatch.toLowerCase() === 'no' && (
-                      <span
-                        title="Live nameservers do not match Cloudflare"
-                        className="px-1.5 py-0.5 rounded text-[11px] font-semibold bg-red-100 text-red-700 border border-red-200"
-                      >
-                        <span aria-hidden="true">⚠ </span>NS
-                      </span>
-                    )}
-                    {s.changed && (
-                      <span
-                        title={s.changed}
-                        className="px-1.5 py-0.5 rounded text-[11px] font-semibold bg-purple-100 text-purple-700 border border-purple-200"
-                      >
-                        Δ
-                        <span className="sr-only">{` Changed since last refresh: ${s.changed}`}</span>
-                      </span>
-                    )}
-                  </div>
-                  <RowActions s={s} />
-                </td>
-                {showRepoCols && (
-                  <>
-                    <td className="px-4 py-2 whitespace-nowrap text-xs">
-                      {s.repoUrl ? (
-                        <a
-                          href={s.repoUrl}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-indigo-600 hover:underline"
-                          title={repoName(s.repoUrl)}
-                        >
-                          {repoName(s.repoUrl).split('/').pop()}
-                        </a>
-                      ) : (
-                        dash
-                      )}
-                    </td>
-                    <td className="px-4 py-2 whitespace-nowrap text-xs text-gray-600">
-                      {s.lastPrClosed || dash}
-                    </td>
-                    <td className="px-4 py-2 whitespace-nowrap text-xs text-gray-600">
-                      {s.openPrs && s.openPrs !== '0' ? s.openPrs : dash}
-                    </td>
-                  </>
-                )}
-                <td className="px-4 py-2 whitespace-nowrap">
-                  <span
-                    className={`px-2 py-0.5 rounded-full text-xs font-semibold border ${healthBadge(s.siteHealth)}`}
-                  >
-                    {s.siteHealth || 'N/A'}
-                  </span>
-                  {s.redirectTarget && (
-                    <p
-                      className="mt-0.5 text-[11px] text-gray-500 max-w-[12rem] truncate"
-                      title={`Redirects to ${s.redirectTarget}`}
-                    >
-                      → {s.redirectTarget}
-                    </p>
-                  )}
-                </td>
-                {hasSsl && (
-                  <td className="px-4 py-2 whitespace-nowrap text-xs">
-                    {s.sslExpiry ? (
-                      <span
-                        className={`px-2 py-0.5 rounded-full text-xs font-semibold border ${sslBadge(s.sslExpiry)}`}
-                        title={`TLS certificate expires ${s.sslExpiry}`}
-                      >
-                        {s.sslExpiry}
-                      </span>
-                    ) : (
-                      dash
-                    )}
-                  </td>
-                )}
-                {hasLh && (
-                  <td className="px-4 py-2 whitespace-nowrap text-xs">
-                    {s.lighthouse ? (
-                      <span
-                        className={`px-2 py-0.5 rounded-full text-xs font-semibold border ${lighthouseBadge(Number(s.lighthouse))}`}
-                        title={`Lighthouse score ${s.lighthouse}/100`}
-                      >
-                        {s.lighthouse}
-                      </span>
-                    ) : (
-                      dash
-                    )}
-                  </td>
-                )}
-                {hasOwner && (
-                  <td className="px-4 py-2 whitespace-nowrap text-xs">
-                    {s.owner ? (
-                      isGithubHandle(s.owner) ? (
-                        <a
-                          href={`https://github.com/${s.owner}`}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-indigo-600 hover:underline"
-                        >
-                          @{s.owner}
-                        </a>
-                      ) : (
-                        s.owner
-                      )
-                    ) : (
-                      dash
-                    )}
-                  </td>
-                )}
-                {showProgress && (
-                  <td className="px-4 py-2 whitespace-nowrap text-xs">
-                    <MigrationProgress s={s} />
-                  </td>
-                )}
-                <td className="px-4 py-2 whitespace-nowrap text-xs text-gray-700">
-                  {s.serverInUse || dash}
-                </td>
-                <td className="px-4 py-2 whitespace-nowrap text-xs text-gray-700">{s.status}</td>
-                <td className="px-4 py-2 text-xs text-gray-500 max-w-xs truncate" title={s.notes}>
-                  {s.notes}
-                </td>
-              </tr>
-            )
-          })}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
 export default function SitesListPage() {
   const sites = loadSites()
   const domainExpiry = loadDomainExpiry()
   const csvUpdated = dataRefreshedAge()
 
-  const byTier = (num: string) => sites.filter((s) => s.workTier.startsWith(num))
-  const counts = Object.fromEntries(TIERS.map((t) => [t.num, byTier(t.num).length]))
-
-  // Domains transferred away from eNom AND no longer in FFC Cloudflare — i.e.
-  // they actually left FFC. Surfaced prominently for cleanup/outreach.
-  const leftFfcSites = sites
-    .filter((s) => s.leftFfc === 'Yes')
-    .sort((a, b) => a.domain.localeCompare(b.domain))
+  const byTier = (num: string) => sites.filter((s) => s.workTier.startsWith(num)).length
 
   const stat = (label: string, value: number, color: string, icon: string) => (
     <div className={`bg-white rounded-lg shadow-md border p-4 text-center ${color}`}>
@@ -445,52 +87,14 @@ export default function SitesListPage() {
 
         {/* Tier stat cards */}
         <div className="grid grid-cols-2 lg:grid-cols-4 gap-4 mb-8">
-          {stat('Active Development', counts['1'], 'border-green-200', '🔨')}
-          {stat('Stalled Repos', counts['2'], 'border-yellow-200', '🌱')}
-          {stat('Needs Migration', counts['3'], 'border-blue-200', '🚚')}
-          {stat('Done / Stable', counts['4'], 'border-teal-200', '✅')}
+          {stat('Active Development', byTier('1'), 'border-green-200', '🔨')}
+          {stat('Stalled Repos', byTier('2'), 'border-yellow-200', '🌱')}
+          {stat('Needs Migration', byTier('3'), 'border-blue-200', '🚚')}
+          {stat('Done / Stable', byTier('4'), 'border-teal-200', '✅')}
         </div>
 
-        {/* Domains that left FFC (transferred away + not in FFC Cloudflare) */}
-        {leftFfcSites.length > 0 && (
-          <details className="mb-8 rounded-lg shadow-md border border-red-300 bg-red-50 overflow-hidden">
-            <summary className="px-6 py-4 cursor-pointer bg-red-100 text-red-900 border-b border-red-200">
-              <span className="text-lg font-bold">🚨 Left FFC — needs attention</span>
-              <span className="ml-2 text-sm font-semibold opacity-80">({leftFfcSites.length})</span>
-              <p className="text-sm mt-1 opacity-80 font-normal">
-                Transferred away from eNom <strong>and</strong> no longer in FFC Cloudflare — FFC no
-                longer manages DNS for these. Still in WHMCS, so worth a cleanup/outreach pass (some
-                are still live under their new owner).
-              </p>
-            </summary>
-            <TierTable sites={leftFfcSites} num="6" />
-          </details>
-        )}
-
-        {/* Work Tier sections */}
-        {TIERS.map((t) => {
-          const tierSites = sortTier(byTier(t.num), t.num)
-          return (
-            <details
-              key={t.num}
-              open={t.open}
-              className="mb-6 rounded-lg shadow-md border border-gray-200 bg-white overflow-hidden"
-            >
-              <summary className={`px-6 py-4 cursor-pointer border-b ${t.header}`}>
-                <span className="text-lg font-bold">
-                  {t.icon} Tier {t.num}: {t.title}
-                </span>
-                <span className="ml-2 text-sm font-semibold opacity-80">({tierSites.length})</span>
-                <p className="text-sm mt-1 opacity-80 font-normal">{t.blurb}</p>
-              </summary>
-              {tierSites.length > 0 ? (
-                <TierTable sites={tierSites} num={t.num} />
-              ) : (
-                <p className="px-6 py-4 text-sm text-gray-400 italic">No domains in this tier.</p>
-              )}
-            </details>
-          )
-        })}
+        {/* Filterable tier tables (#412–#415, #426) */}
+        <SitesExplorer sites={sites} />
 
         {/* Data provenance legend (#417) */}
         <details className="mb-6 rounded-lg border border-gray-200 bg-white px-6 py-4 text-sm text-gray-700 shadow-md">
